Add linter for error catching statements in tests
Summary:
This diff adds a linter for exception catching statements
like below in tests:

```lang=scala
an [Exception] shouldBe thrownBy stmt
a [RuntimeError] shouldBe thrownBy stmt
the [Exception] thrownBy stmt
```

The linter ensures that the correct article between
a/an is being used. In addition, it also ensures that
there's exactly one space between the article and the
opening square bracket.

Test Plan:
 - unit tests
 - screenshot: {F65458}

Reviewers: hao, arijit

Reviewed By: hao

Differential Revision: https://phabricator.rubrik.com/D32843