Make all lint issues into warnings
Summary:
If the severity of a lint message is "error", then
arcanist shows the lint message irrespective of whether
that particular line is touched in the diff or not.

We're not ready for large refactors yet, so in this diff
we set the severity of all lint messages to warning if
either the severity is explicitly mentioned as "error",
or it is not mentioned at all.

Test Plan: manual verification