Add linter to enforce one-space-after-if rule
Summary:
This diff adds a linter that ensures exactly one space
exists between the `if` keywoard and the next opening
parenthesis.

We could've used scalastyle's `EnsureSingleSpaceAfterTokenChecker`,
but it is buggy. For example, it fails to catch the following case:

```lang=scala
object LifeUniverseAndEverything {
  def answer: Int = 42
  if(false) implode()
}
```

I did try to patch scalastyle, but their code is convoluted,
ridden with boolean flags that don't make much sense and
abstractions that simply exist to templatize code rather than
because the abstractions are logical. A major rewrite of
the scalastyle class would likely get rejected.

In addition, scalastyle doesn't suggest replacements, which
`arc` can use to auto-patch lint errors. The rule implemented
in this diff has the autofix functionality (see screenshot in
the test plan).

Test Plan:
 - Newly added unit tests
 - Screenshot of `arc lint`: {F67201}

Reviewers: mudit, hao

Reviewed By: mudit

Differential Revision: https://phabricator.rubrik.com/D33269