Update call-chain linter rules
Summary:
This diff updates the call-chain linter rules to be the following:

1) Either all the members of the call-chain are on the same line
2) Or, if a line break is needed, then at least all the members
   after and including the first method call are on a line of their
   own and left-aligned.

That is, a prefix of the chain is allowed to be on the same line
given that all the chain members in that prefix are property
accesses (or, at least, look like property accesses) and not method
calls. And after that, every chain member gets a line of its own.

We modified this rule to allow use cases like the following:

```lang=scala
deps.cluster.vcd.api
  .bigFunctionName(
    arg1,
    arg2)
  .map(...)
```

Test Plan:
 - Existing test cases
 - Newly added test case

Reviewers: jamie

Reviewed By: jamie

Differential Revision: https://phabricator.rubrik.com/D33541