Ban `should be`s that could possibly dangle
Summary:
The following code compiles,
but it doesn't do what it is expected to do:
```lang=scala
resultVal should be
(expectedVal)
// passes even if resultVal is not expectedVal
```

The code above is not equivalent to
```lang=scala
resultVal should be (expectedVal)
```

The first one compiles because `foo.should(be)` is valid in
scalatest in order to enable syntax like the following:
```lang=scala
foo should be a bar
// same as
foo.should(be).a(bar)
```

The convenience brought by the example above
is not worth it especially given how easily
example (1) could go unnoticed in reviews. Also, given that
`foo.should(be)` can always be replaced by `foo.shouldBe`,
by banning the usage of the former, we aren't making any
test constructs impossible to be written.

Test Plan:
 - Newly added unit tests
 - Manually tested on `sdmain`

Reviewers: jamie, jd, sat, shaswat.chaubey, mudit

Reviewed By: jd

Differential Revision: https://phabricator.rubrik.com/D34544