Add rule to warn on `new Date[Time]`
Summary:
We want our code in general, and tests in particular,
to not depend on the particular date/time at the time
of execution. Also, we want to be able to control both
the time instances and the process of advancement of time
in our tests. It is not possible to do that if production
code directly created time instances using `new Date` or
`new DateTime`.

Instead, the code should use a `Clock` object, which can
be injected and then we can use different clock objects
in production (one that uses the system clock) and
in tests (one that can be set and manually advanced).

This diff adds a linter that catches instances where
new date / datetime objects are being created
to get hold of current time and raises warnings.

Test Plan: - Newly added unit tests

Reviewers: jamie

Reviewed By: jamie

Differential Revision: https://phabricator.rubrik.com/D34717