Warn for 'should not be' and suggest 'equal'
Summary:
This diff adds a linter that warns against using
the `... should not be ...` pattern in tests and
suggests usage of the `... should not equal ...`
pattern instead.

`should not be` is problematic because the
following code compiles, and the would pass too,
but the test isn't testing what really should
be tested.

```lang=scala
def getWrongAnswer(): Option[Int] = {
  Some(42)
  // Whoops!
  // we were supposed to return a wrong answer!
  // but look! we're actually returning 42!!
}

getWrongAnswer() should not be 42
// Observe that the test author feels like they're
// actually testing that 42 isn't returned, but ah
// the tragedy of types! The test passes because where
// an actual type-error should've been raised, it wasn't.
```

`should not equal` on the other hand, is type-safe
and the following code simply won't compile, and we
will get an error instead telling us that return type
of `getWrongAnswer()` doesn't match that of `42`

```lang=scala
getWrongAnswer() should not equal 42
```

Test Plan: - Newly added unit test

Reviewers: sat

Reviewed By: sat

Differential Revision: https://phabricator.rubrik.com/D42869