Support linting multiple files
Summary:
Starting the JVM is the biggest bottleneck for `sclinter`
right now. Invoking `sclinter` each time for every file
to be linted is thus extremely inefficient. This diff is
the first step toward improving the linting speed.

In a future diff, we will change the way `sclinter` is
invoked from `arc lint`.

Test Plan:
 - Introduced lint errors in multiple files and ensured
   that they are caught, thus making sure that there's no
   regression.
 - Made hacky changes to the arcanist linter to test for
   speedup. `arc lint --everything` went from 40 seconds
   to 5 seconds.
 - No further testing as this diff doesn't change any
   functionality.

Reviewers: arijit

Reviewed By: arijit

Subscribers: hao

Differential Revision: https://phabricator.rubrik.com/D43064