Remove submodule, maintain local copy
Summary:
To make `arc lint` faster with `sclinter`, as the
first step, D43064 made `sclinter` multi-file aware
so that there is no need to invoke `sclinter` once
for each file, but rather, we can simply lint all
the files with a single invocation.

However, `ArcanistExternalJsonLinter` does not support
passing multiple files as arguments to the underlying
script, and hence needs modification. Before we can
modify it, we need to take it out of the git submodule
and add it locally to the repository. This diff simply
gets rid of the submodule.

Test Plan:
 - Make sure there are no regressions by manually
   introducing multiple lint errors in scala files and
   making sure that `arc lint` successfully catches them
   all.

Reviewers: arijit, Alfred

Reviewed By: arijit

Differential Revision: https://phabricator.rubrik.com/D43066