Invoke sclinter only once per arc lint
Summary:
D43064 made it possible to invoke `sclinter` only
once to lint multiple files. This diff modifies
`ArcanistExternalJsonLinter` so that it passes
the list of files to be invoked to only one instance
of `sclinter` instead of spawning one instance
per file.

Obsevred speedup from 40 seconds to 5 seconds
for `arc lint --everything` on the `sclinter` repo.

Test Plan:
 - Manually verify that lint errors are caught in multiple
   files appropriately

Reviewers: arijit, Alfred

Reviewed By: arijit

Differential Revision: https://phabricator.rubrik.com/D43067