Compile to JS too
Summary:
We eventually want to be able to use sclinter
functionality from python (python doesn't have
a scala parser). As a first step, in this diff
we cross-compile most of `sclinter` (except the
main app) to JS.

In future diffs, JS classes will be exported and
transpiled to python.

Test Plan:  - Existing unit tests

Reviewers: arijit, Alfred

Reviewed By: arijit

Differential Revision: https://phabricator.rubrik.com/D43260