Add linter to catch variable shadowing
Summary:
This diff adds a `Linter` that catches instances when a name in an
inner scope shadows a name in any of its enclosing scopes.

Limitations:
Just like all the other `Linter`s based on `org.scalameta`,
is `ShadowLinter` is purely a syntax-based linter that is
not type aware, whenever the shadowed name is not explicitly
present in the outer scope, such instances will not be caught
by `ShadowLinter`. For example, shadowing a name that is
imported through a wildcard import, or shadowing names that are
part of an outer scope through inheritance.

Test Plan: - Newly added unit tests

Reviewers: hao, sat, nishant.gupta

Reviewed By: sat, nishant.gupta

Subscribers: nishant.gupta

Differential Revision: https://phabricator.rubrik.com/D44285